keybinds options sections and cleanup
